lets find the diffs

+++ src/utils/populateDb.js
@@ -2,20 +2,28 @@
 /* eslint-disable security/detect-eval-with-expression */
 const fs = require('fs');
 const path = require('path');
-
+const { glob } = require('glob');
 const logger = require('../config/logger');
 const Sheetdefinition = require('../models/sheetdefinition.model');
 
+async function globby() {
+  const files = await glob('**/*', { cwd: 'assets/spritesheets', ignore: 'node_modules' });
+  let renamed = [];
+  renamed = files.map((f) => f.replaceAll('/', '_'));
+  renamed.forEach((f) => logger.info(f));
+  return renamed;
+}
+
+globby();
+
 // Function to check if the collection is empty and populate it if needed
 async function checkAndPopulateCollection() {
-  Sheetdefinition.collection.drop();
+  if (Sheetdefinition.collection.length > 0) {
+    Sheetdefinition.collection.drop();
+  }
   try {
-    // const count = await SheetDefinition.countDocuments({});
-
-    // fires everytime now for debugging purposes
-    // if (count || !count)
-
     // Define your default JSON data
+
     const sheetDefinitionDirectory = path.join(__dirname, '../../assets/sheet_definitions');
     const files = fs.readdirSync(sheetDefinitionDirectory);
